+++ src/main.rs
@@ -1,8 +1,5 @@
-use sqlx::postgres::PgPoolOptions;
-use std::net::TcpListener;
 use zero2prod::configuration::get_configuration;
-use zero2prod::email_client::EmailClient;
-use zero2prod::startup::run;
+use zero2prod::startup::Application;
 use zero2prod::telemetry::{get_subscriber, init_subscriber};
 
 #[tokio::main]
@@ -11,28 +8,7 @@ async fn main() -> std::io::Result<()> {
     init_subscriber(subscriber);
     let configuration = get_configuration().expect("Failed to read configuration");
 
-    let connection_pool = PgPoolOptions::new()
-        .connect_timeout(std::time::Duration::from_secs(2))
-        .connect_lazy_with(configuration.database.with_db());
-
-    let sender_email = configuration
-        .email_client
-        .sender()
-        .expect("Invalid sender email address.");
-
-    let timeout = configuration.email_client.timeout();
-    let email_client = EmailClient::new(
-        configuration.email_client.base_url,
-        sender_email,
-        configuration.email_client.auth_token,
-        timeout,
-    );
-
-    let port = std::env::var("PORT").unwrap_or_else(|_| configuration.application.port.to_string());
-
-    let address = format!("{}:{}", configuration.application.host, port);
-    tracing::info!("Address: {:?}", address);
-    let listener = TcpListener::bind(address).expect("Failed to bind random port");
-
-    run(listener, connection_pool, email_client)?.await
+    let application = Application::build(configuration.clone()).await?;
+    application.run().await?;
+    Ok(())
 }
